project-avatar-api/requirements.txt
# new_avatar\projects-avatar-api\requirements.txt

annotated-types==0.7.0       
anyio==4.8.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
mysqlclient==2.2.7
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.7.1     
pydantic_core==2.27.2        
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
pydantic[email]==2.10.6
pymysql==1.1.1